legacy/pyproject.toml
[project]
name = "celestial-explorer"
version = "0.1.0"
description = "A Google Maps-like interface for exploring space through satellite perspectives"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "astropy>=6.1.7",
    "dash>=3.2.0",
    "pandas>=2.3.3",
    "requests>=2.31.0",
    "plotly>=5.17.0",
    "numpy>=1.24.0",
    "lxml>=4.9.0",
    "pillow>=10.0.0",
]
